Precompile the normal_vector folder
