+++ include/meltpooldg/normal_vector/normal_vector_operation_base.hpp
@@ -27,15 +27,7 @@ namespace MeltPoolDG
                  const Parameters<double> &                     data_in,
                  const unsigned int                             normal_dof_idx_in,
                  const unsigned int                             normal_quad_idx_in,
-                 const unsigned int                             ls_dof_idx_in)
-      {
-        (void)scratch_data_in;
-        (void)data_in;
-        (void)normal_dof_idx_in;
-        (void)normal_quad_idx_in;
-        (void)ls_dof_idx_in;
-        AssertThrow(false, ExcNotImplemented());
-      }
+                 const unsigned int                             ls_dof_idx_in);
 
       virtual void
       reinit() = 0;
